refactor: replace NotebookTree with context bar version
